+++ pkg/unittest/assertion.go
@@ -50,8 +50,13 @@ func (a *Assertion) Assert(
 		}
 
 		if a.requireRenderSuccess != renderSucceed {
-			invalidRender := []string{fmt.Sprintf("Error: Invalid rendering: %s", rendered[0][common.RAW])}
-			failInfo = append(failInfo, invalidRender...)
+			invalidRender := ""
+			if len(rendered) > 0 {
+				invalidRender = fmt.Sprintf("Error: Invalid rendering: %s", rendered[0][common.RAW])
+			} else {
+				invalidRender = "Error: rendered manifest is empty"
+			}
+			failInfo = append(failInfo, invalidRender)
 			break
 		}
 
@@ -187,4 +192,5 @@ var assertTypeMapping = map[string]assertTypeDef{
 	"failedTemplate":    {reflect.TypeOf(validators.FailedTemplateValidator{}), false, false},
 	"notFailedTemplate": {reflect.TypeOf(validators.FailedTemplateValidator{}), true, true},
 	"containsDocument":  {reflect.TypeOf(validators.ContainsDocumentValidator{}), false, true},
+	"lengthEqual":       {reflect.TypeOf(validators.LengthEqualDocumentsValidator{}), false, true},
 }