+++ pkg/unittest/validators/contains_document_validator.go
@@ -3,6 +3,7 @@ package validators
 import (
 	"fmt"
 
+	"github.com/lrills/helm-unittest/internal/common"
 	"github.com/lrills/helm-unittest/pkg/unittest/valueutils"
 )
 
@@ -15,11 +16,50 @@ type ContainsDocumentValidator struct {
 }
 
 func (v ContainsDocumentValidator) failInfo(actual interface{}, index int, not bool) []string {
+
 	return splitInfof(
 		setFailFormat(not, false, false, false, " to contain document"),
 		index,
 		fmt.Sprintf("Kind = %s, apiVersion = %s", v.Kind, v.APIVersion),
 	)
+}
+
+func (v ContainsDocumentValidator) validateManifest(manifest common.K8sManifest, negative bool) bool {
+	if kind, ok := manifest["kind"].(string); (ok && kind == v.Kind) == negative {
+		// if no match, move onto next document
+		return false
+	}
+
+	if api, ok := manifest["apiVersion"].(string); (ok && api == v.APIVersion) == negative {
+		// if no match, move onto next document
+		return false
+	}
+
+	if v.Name != "" {
+		actual, err := valueutils.GetValueOfSetPath(manifest, "metadata.name")
+		if err != nil {
+			// fail on not found match
+			return false
+		}
+
+		if (actual == v.Name) == negative {
+			return false
+		}
+	}
+
+	if v.Namespace != "" {
+		actual, err := valueutils.GetValueOfSetPath(manifest, "metadata.namespace")
+		if err != nil {
+			// fail on not found match
+			return false
+		}
+
+		if (actual == v.Namespace) == negative {
+			return false
+		}
+	}
+
+	return true
 }
 
 // Validate implement Validatable
@@ -33,47 +73,15 @@ func (v ContainsDocumentValidator) Validate(context *ValidateContext) (bool, []s
 	validateErrors := make([]string, 0)
 
 	for _, manifest := range manifests {
-		if kind, ok := manifest["kind"].(string); (ok && kind == v.Kind) == context.Negative {
-			// if no match, move onto next document
-			continue
+		validateSuccess = v.validateManifest(manifest, context.Negative)
+		if validateSuccess {
+			break
 		}
-
-		if api, ok := manifest["apiVersion"].(string); (ok && api == v.APIVersion) == context.Negative {
-			// if no match, move onto next document
-			continue
-		}
-
-		if v.Name != "" {
-			actual, err := valueutils.GetValueOfSetPath(manifest, "metadata.name")
-			if err != nil {
-				// fail on not found match
-				continue
-			}
-
-			if (actual == v.Name) == context.Negative {
-				continue
-			}
-		}
-
-		if v.Namespace != "" {
-			actual, err := valueutils.GetValueOfSetPath(manifest, "metadata.namespace")
-			if err != nil {
-				// fail on not found match
-				continue
-			}
-
-			if (actual == v.Namespace) == context.Negative {
-				continue
-			}
-		}
-
-		// if we get here the above have held so it is a match
-		validateSuccess = true
-		break
+		continue
 	}
 	if !validateSuccess {
-		errorMesasge := v.failInfo(v.Kind, 0, context.Negative)
-		validateErrors = append(validateErrors, errorMesasge...)
+		errorMessage := v.failInfo(v.Kind, 0, context.Negative)
+		validateErrors = append(validateErrors, errorMessage...)
 	}
 	validateSuccess = determineSuccess(1, validateSuccess, true)